Rename to filtered positive identifier
